+++ nuxt.config.js
@@ -1,27 +1,16 @@
-require('dotenv').config(); // parse .env file
-
 export default {
-  build: {
-    maxChunkSize: 300000, // split into smaller files
-    postcss: {
-      plugins: {
-        'postcss-custom-properties': {
-          // see https://github.com/buefy/buefy/issues/306
-          // and https://github.com/nuxt/nuxt.js/issues/1670
-          warnings: false
-        }
-      }
-    }
-  },
-
   buildModules: ['@nuxtjs/style-resources'],
-  components: true, // auto import components
+
+  // auto import components
+  components: true,
+
   css: ['assets/scss/main'],
 
   head: {
     htmlAttrs: {
       lang: 'en'
     },
+
     link: [
       {
         color: '#36d1c4',
@@ -62,7 +51,9 @@ export default {
         rel: 'manifest'
       }
     ],
-    meta: [ // charset and viewport should come first
+
+    meta: [
+      // charset and viewport should come first
       {
         charset: 'utf-8'
       },
@@ -95,50 +86,80 @@ export default {
         name: 'robots'
       },
       {
-        content: 'yes',
-        name: 'mobile-web-app-capable'
-      },
-      {
-        content: 'peerhaven',
+        content: 'Peerhaven',
         name: 'apple-mobile-web-app-title'
       },
       {
-        content: 'peerhaven',
+        content: 'Peerhaven',
         name: 'application-name'
       },
       {
-        content: 'peerhaven is a place for you to save and share your bookmarks.',
+        content: 'Peerhaven is a place for you to save and share your bookmarks.',
         name: 'description'
+      },
+      {
+        content: 'telephone=no',
+        name: 'format-detection'
+      },
+      {
+        content: 'yes',
+        name: 'mobile-web-app-capable'
       }
     ],
-    title: 'peerhaven'
+
+    title: 'Peerhaven'
   },
 
-  loading: { // custom progress bar
-    color: '#f6318c',
+  // custom progress bar
+  loading: {
+    color: 'hsl(332, 92%, 58%)',
     continuous: true,
-    failedColor: '#ff3860'
+    failedColor: 'hsl(348, 100%, 61%)'
   },
 
   modules: [
     [
       'nuxt-buefy',
       {
-        css: false // use SCSS imports instead
-      }
-    ],
-    [
-      'nuxt-matomo',
-      {
-        cookies: true,
-        matomoUrl: process.env.MATOMO_URL,
-        siteId: process.env.MATOMO_SITE_ID
+        // use SCSS imports instead
+        css: false,
+
+        materialDesignIcons: false
       }
     ]
   ],
 
   render: {
-    csp: true // see https://github.com/nuxt/nuxt.js/pull/2549
+    csp: {
+      hashAlgorithm: 'sha512',
+
+      policies: {
+        'base-uri': ["'self'"],
+        'child-src': ["'none'"],
+        'connect-src': ["'self'"],
+        'default-src': ["'none'"],
+        'font-src': ['fonts.gstatic.com'],
+        'form-action': ["'none'"],
+        'frame-ancestors': ["'none'"],
+        'frame-src': ["'none'"],
+        'img-src': ["'self'"],
+        'manifest-src': ["'self'"],
+        'media-src': ["'none'"],
+        'object-src': ["'none'"],
+
+        // SHA-2 and 'self' injected by Nuxt
+        'script-src': [],
+
+        'style-src': [
+          // see https://github.com/buefy/nuxt-buefy/issues/97
+          "'unsafe-inline'",
+
+          'fonts.googleapis.com'
+        ],
+
+        'worker-src': ["'none'"]
+      }
+    }
   },
 
   styleResources: {