+++ CHANGELOG.md
@@ -4,16 +4,23 @@
 >
 > (Source: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/))
 
-## 1.0.1
+## 1.0.1 · Unreleased
 
 ### Added
 
+- Console warning to help prevent [social engineering](https://en.wikipedia.org/wiki/Social_engineering_%28security%29) attacks
+- [Content Security Policy](https://content-security-policy.com/)
 - Error pages
+- License: [Fair Source 5 (v0.9)](https://fair.io/)
 
-## 1.0.0
+### Removed
+
+- Analytics by [Matomo](https://matomo.org/)
+
+## 1.0.0 · 2018-04-25
 
 ### Added
 
-- peerhaven logo
+- Peerhaven logo
 - [Favicons](https://en.wikipedia.org/wiki/Favicon) for various platforms (iOS, Android, Windows, browsers)
 - Analytics by [Matomo](https://matomo.org/)